% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the occurence of fatal injuries by state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numeric state number}

\item{year}{A numeric year input}
}
\value{
A map graphic of the selected state.
}
\description{
This function plots the occurence of fatal injuries suffered in motor vehicle traffic crashes
}
\note{
Error may occur if the state input is invalid or if there are no accidents to plot.
}
\examples{
\dontrun{
fars_map_state(1, 2015)
fars_map_state(50, 2013)
}
}
