% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads data and creates a data frame tbl}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of the file that is read into R}
}
\value{
This function will return a data frame of format tbl_df. (For further information look at tbl_df{dplyr})
}
\description{
This function reads a csv file if existant and creates a data frame tbl.
}
\examples{
#'
\dontrun{
fars_read("accident_2015.csv.bz2")
}
}
