% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Creates a list of tibbles}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A list of tibbles
}
\description{
This function returns the observation "month" and "year" for a list of datasets.
}
\note{
Invalid year values may result in an error and NULL is returned.
}
\examples{
fars_read_years(c("2013", "2014"))
}
