% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Shows number of observations for each month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A tibble
}
\description{
This function returns a tibble that summarizes the number of observations for each month and each dataset.
}
\note{
Invalid year values may result in an error and NULL is returned for the specific dataset.
}
\examples{
fars_summarize_years(c("2013", "2014"))
}
