% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Makes filename in C-style Formatting}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year of the document}
}
\value{
This function will return a character-string.
}
\description{
This function produces a filename out of a year input in C-style String formatting.
}
\examples{
make_filename(year = 2019)
}
